refactor(check): split check and CLI modules
