+++ src/cli/src/main.rs
@@ -1,22 +1,21 @@
+mod args;
+mod check_runner;
 mod config;
+mod options;
 
-use crate::config::{
-    ConfigSampleCommand, load_pyproject_config, render_config_sample, resolve_config_path,
+use crate::args::{CheckOutputFormat, Cli, Commands, ConfigCommands, FailSeverity};
+use crate::check_runner::{CheckRunConfig, expand_check_kinds, run_check};
+use crate::config::{load_pyproject_config, render_config_sample};
+use crate::options::{
+    cli_or_config_enum, cli_or_config_enum_vec, cli_or_config_path, cli_or_config_vec,
+    exit_config_error, normalize_output_path, resolve_required_path, write_output_file,
 };
-use anyhow::{Context, Result};
-use check::{
-    CheckCodeConfig, CheckLocaleCache, CheckResult, Diagnostic, DiagnosticKind, Severity,
-    check_kwargs_with_cache, check_missing_with_cache, check_references_with_cache,
-    check_stale_with_cache, check_syntax_with_cache, check_untranslated_with_cache,
-    code_extraction_errors, extract_check_code, has_failing_diagnostics, render_check_json,
-    render_check_terminal, validate_check_locales,
-};
-use clap::{Parser, Subcommand, ValueEnum};
+use check::{has_failing_diagnostics, render_check_json, render_check_terminal};
+use clap::Parser;
 use extractor::ftl::consts::{
     CommentsKeyModes, DEFAULT_EXCLUDE_DIRS, DEFAULT_FTL_FILENAME, DEFAULT_I18N_KEYS,
     DEFAULT_IGNORE_ATTRIBUTES, DEFAULT_IGNORE_KWARGS, LineEndings,
 };
-use extractor::ftl::diagnostics::ExtractedCode;
 use extractor::ftl::ftl_extractor::{ExtractConfig, extract};
 use extractor::ftl::utils::FastHashSet;
 use log::{error, info};
@@ -26,184 +25,6 @@ use stub::{StubConfig, generate_stub};
 
 #[global_allocator]
 static GLOBAL: MiMalloc = MiMalloc;
-
-#[derive(Parser)]
-#[command(name = "ftl", version, about)]
-struct Cli {
-    /// Path to pyproject.toml with [tool.ftl-extract.<command>] config
-    #[arg(long)]
-    config: Option<PathBuf>,
-
-    #[command(subcommand)]
-    command: Option<Commands>,
-
-    /// Verbose output
-    #[arg(short = 'v', long, global = true, default_value_t = false)]
-    verbose: bool,
-}
-
-#[derive(Subcommand)]
-enum Commands {
-    Config {
-        #[command(subcommand)]
-        command: ConfigCommands,
-    },
-    Extract {
-        /// Path to the code directory
-        #[arg()]
-        code_path: Option<PathBuf>,
-
-        /// Path to the output directory
-        #[arg()]
-        output_path: Option<PathBuf>,
-
-        /// Language codes to extract
-        #[arg(short = 'l', long)]
-        language: Vec<String>,
-
-        /// Names of function that is used to get translation
-        #[arg(short = 'k', long)]
-        i18n_keys: Vec<String>,
-
-        /// Append names of function that is used to get translation
-        #[arg(short = 'K', long, default_values_t = Vec::<String>::new())]
-        i18n_keys_append: Vec<String>,
-
-        /// Prefix names of function that is used to get translation. `self.i18n.*()`
-        #[arg(short = 'p', long, default_values_t = Vec::<String>::new())]
-        i18n_keys_prefix: Vec<String>,
-
-        /// Exclude directories
-        #[arg(short = 'e', long)]
-        exclude_dirs: Vec<String>,
-
-        /// Append directories to exclude
-        #[arg(short = 'E', long, default_values_t = Vec::<String>::new())]
-        exclude_dirs_append: Vec<String>,
-
-        /// Ignore attributes, e.g. `i18n.set_locale()`
-        #[arg(short = 'i', long)]
-        ignore_attributes: Vec<String>,
-
-        /// Append attributes to ignore
-        #[arg(short = 'I', long, default_values_t = Vec::<String>::new())]
-        append_ignore_attributes: Vec<String>,
-
-        /// Ignore kwargs, like `when` from `aiogram_dialog.I18nFormat(..., when=...)`
-        #[arg(long)]
-        ignore_kwargs: Vec<String>,
-
-        /// Comment Junk elements
-        #[arg(long, default_value_t = false)]
-        comment_junks: bool,
-
-        /// Default FTL filename
-        #[arg(long)]
-        default_ftl_file: Option<PathBuf>,
-
-        /// Comment keys mode
-        #[arg(long, value_enum)]
-        comment_keys_mode: Option<CommentsKeyModes>,
-
-        /// Line endings in output FTL files
-        #[arg(long, value_enum)]
-        line_endings: Option<LineEndings>,
-
-        /// Dry run, do not write to files
-        #[arg(long, default_value_t = false)]
-        dry_run: bool,
-
-        /// Cache Python extraction results between runs
-        #[arg(long, default_value_t = false)]
-        cache: bool,
-
-        /// Directory or file path for the extraction cache
-        #[arg(long)]
-        cache_path: Option<PathBuf>,
-
-        /// Clear the extraction cache before running
-        #[arg(long, default_value_t = false)]
-        clear_cache: bool,
-    },
-    Stub {
-        /// Path to the FTL files directory
-        #[arg()]
-        ftl_path: Option<PathBuf>,
-
-        /// Output path for the .pyi stub file
-        #[arg()]
-        output_path: Option<PathBuf>,
-
-        /// Export intermediate tree structure as JSON
-        #[arg(long, default_value_t = false)]
-        export_tree: bool,
-    },
-    Check {
-        /// Path to locales directory containing locale folders (e.g. en, uk)
-        #[arg()]
-        locales_path: Option<PathBuf>,
-
-        /// Path to Python code for checks that inspect source usage
-        #[arg(long)]
-        code_path: Option<PathBuf>,
-
-        /// Checks to run
-        #[arg(long = "check", value_enum)]
-        checks: Vec<CheckKind>,
-
-        /// Locale codes to check
-        #[arg(short = 'l', long = "language", default_values_t = Vec::<String>::new())]
-        language: Vec<String>,
-
-        /// Suggest translations from these locales
-        #[arg(long, default_values_t = Vec::<String>::new())]
-        suggest_from: Vec<String>,
-
-        /// Minimum diagnostic severity that should fail the command
-        #[arg(long, value_enum, default_values_t = Vec::<FailSeverity>::new())]
-        fail_on: Vec<FailSeverity>,
-
-        /// Output report path for batch processing
-        #[arg(long)]
-        output: Option<PathBuf>,
-
-        /// Output format for report file
-        #[arg(long, value_enum)]
-        output_format: Option<CheckOutputFormat>,
-    },
-}
-
-#[derive(Subcommand)]
-enum ConfigCommands {
-    Sample {
-        /// Print only one command-specific pyproject.toml section
-        #[arg(long, value_enum)]
-        command: Option<ConfigSampleCommand>,
-    },
-}
-
-#[derive(PartialEq, Eq, Clone, Debug, clap::ValueEnum)]
-enum CheckKind {
-    All,
-    Kwargs,
-    Missing,
-    References,
-    Stale,
-    Syntax,
-    Untranslated,
-}
-
-#[derive(PartialEq, Clone, Debug, clap::ValueEnum)]
-enum CheckOutputFormat {
-    Terminal,
-    Json,
-}
-
-#[derive(PartialEq, Clone, Debug, clap::ValueEnum)]
-enum FailSeverity {
-    Error,
-    Warn,
-}
 
 fn main() {
     let cli = Cli::parse();
@@ -612,535 +433,5 @@ fn main() {
 
     if let Some(elapsed) = elapsed {
         info!(target: "cli", "✅ Done in {:.3?}s.", elapsed.as_secs_f64());
-    }
-}
-
-fn write_output_file(path: &Path, content: String) -> std::io::Result<()> {
-    if let Some(parent) = path.parent()
-        && !parent.as_os_str().is_empty()
-    {
-        std::fs::create_dir_all(parent)?;
-    }
-    std::fs::write(path, content)
-}
-
-#[derive(Clone)]
-struct CheckRunConfig {
-    locales_path: PathBuf,
-    code_path: Option<PathBuf>,
-    locales: Vec<String>,
-    suggest_from: Vec<String>,
-    i18n_keys: FastHashSet<String>,
-    i18n_keys_prefix: FastHashSet<String>,
-    exclude_dirs: FastHashSet<String>,
-    ignore_attributes: FastHashSet<String>,
-    ignore_kwargs: FastHashSet<String>,
-    default_ftl_file: PathBuf,
-    cache: bool,
-    cache_path: Option<PathBuf>,
-    clear_cache: bool,
-}
-
-fn run_check(expanded_checks: ExpandedChecks, config: CheckRunConfig) -> Result<CheckResult> {
-    run_check_with_extractor(expanded_checks, config, extract_check_code)
-}
-
-fn run_check_with_extractor<F>(
-    expanded_checks: ExpandedChecks,
-    config: CheckRunConfig,
-    mut extract_code: F,
-) -> Result<CheckResult>
-where
-    F: FnMut(CheckCodeConfig) -> Result<ExtractedCode>,
-{
-    let mut result = CheckResult {
-        checked_kinds: Vec::new(),
-        diagnostics: Vec::new(),
-    };
-    let mut extracted_code = None;
-    let mut locale_cache = None;
-    let mut extraction_diagnostics_added = false;
-
-    for check in expanded_checks.checks {
-        match check {
-            CheckKind::All => unreachable!("check expansion removes `all`"),
-            CheckKind::Kwargs => {
-                let extracted =
-                    ensure_extracted_code(&mut extracted_code, &config, &mut extract_code)?;
-                let cache = ensure_locale_cache(&mut locale_cache, &config, &[])?;
-                extend_check_result(
-                    &mut result,
-                    CheckResult::from(check_kwargs_with_cache(cache, extracted)?),
-                );
-                add_extraction_diagnostics_once(
-                    &mut result,
-                    extracted,
-                    &mut extraction_diagnostics_added,
-                );
-            }
-            CheckKind::Missing => {
-                let extracted =
-                    ensure_extracted_code(&mut extracted_code, &config, &mut extract_code)?;
-                let cache = ensure_locale_cache(&mut locale_cache, &config, &[])?;
-                extend_check_result(
-                    &mut result,
-                    CheckResult::from(check_missing_with_cache(cache, extracted)?),
-                );
-                add_extraction_diagnostics_once(
-                    &mut result,
-                    extracted,
-                    &mut extraction_diagnostics_added,
-                );
-            }
-            CheckKind::References => {
-                let cache = ensure_locale_cache(&mut locale_cache, &config, &[])?;
-                extend_check_result(
-                    &mut result,
-                    CheckResult::from(check_references_with_cache(cache)?),
-                );
-            }
-            CheckKind::Stale => {
-                let extracted =
-                    ensure_extracted_code(&mut extracted_code, &config, &mut extract_code)?;
-                let cache = ensure_locale_cache(&mut locale_cache, &config, &[])?;
-                extend_check_result(
-                    &mut result,
-                    CheckResult::from(check_stale_with_cache(cache, extracted)?),
-                );
-                add_extraction_diagnostics_once(
-                    &mut result,
-                    extracted,
-                    &mut extraction_diagnostics_added,
-                );
-            }
-            CheckKind::Syntax => {
-                let cache = ensure_locale_cache(&mut locale_cache, &config, &[])?;
-                extend_check_result(
-                    &mut result,
-                    CheckResult::from(check_syntax_with_cache(cache)?),
-                );
-                if has_fatal_syntax_diagnostics(&result) {
-                    break;
-                }
-            }
-            CheckKind::Untranslated => {
-                let cache = ensure_locale_cache(&mut locale_cache, &config, &config.suggest_from)?;
-                extend_check_result(
-                    &mut result,
-                    CheckResult::from(check_untranslated_with_cache(cache, &config.suggest_from)?),
-                );
-            }
-        }
-    }
-
-    dedup_diagnostics(&mut result);
-
-    Ok(result)
-}
-
-fn ensure_locale_cache<'a>(
-    locale_cache: &'a mut Option<CheckLocaleCache>,
-    config: &CheckRunConfig,
-    extra_locales: &[String],
-) -> Result<&'a CheckLocaleCache> {
-    if locale_cache.is_none() {
-        *locale_cache = Some(CheckLocaleCache::load(
-            &config.locales_path,
-            &config.locales,
-            extra_locales,
-        )?);
-    } else if !extra_locales.is_empty()
-        && let Some(cache) = locale_cache.as_mut()
-    {
-        cache.load_extra_locales(extra_locales)?;
-    }
-
-    Ok(locale_cache.as_ref().expect("locale cache is initialized"))
-}
-
-fn ensure_extracted_code<'a, F>(
-    extracted_code: &'a mut Option<ExtractedCode>,
-    config: &CheckRunConfig,
-    extract_code: &mut F,
-) -> Result<&'a ExtractedCode>
-where
-    F: FnMut(CheckCodeConfig) -> Result<ExtractedCode>,
-{
-    if extracted_code.is_none() {
-        validate_check_locales(&config.locales_path, &config.locales)?;
-        let code_path = config.code_path.clone().context(
-            "Missing code path. Pass --code-path or set tool.ftl-extract.check.code-path",
-        )?;
-        *extracted_code = Some(extract_code(CheckCodeConfig {
-            code_path,
-            i18n_keys: config.i18n_keys.clone(),
-            i18n_keys_prefix: config.i18n_keys_prefix.clone(),
-            exclude_dirs: config.exclude_dirs.clone(),
-            ignore_attributes: config.ignore_attributes.clone(),
-            ignore_kwargs: config.ignore_kwargs.clone(),
-            default_ftl_file: config.default_ftl_file.clone(),
-            cache: config.cache,
-            cache_path: config.cache_path.clone(),
-            clear_cache: config.clear_cache,
-        })?);
-    }
-
-    Ok(extracted_code
-        .as_ref()
-        .expect("extracted code is initialized"))
-}
-
-fn add_extraction_diagnostics_once(
-    result: &mut CheckResult,
-    extracted: &ExtractedCode,
-    added: &mut bool,
-) {
-    if *added {
-        return;
-    }
-
-    result
-        .diagnostics
-        .extend(
-            code_extraction_errors(extracted)
-                .into_iter()
-                .map(|item| Diagnostic {
-                    severity: Severity::Error,
-                    kind: DiagnosticKind::Extraction,
-                    locale: None,
-                    key: Some(item.key),
-                    ftl_location: None,
-                    code_location: item.locations.first().cloned(),
-                    message: item.message,
-                    suggestions: Vec::new(),
-                    missing_kwargs: Vec::new(),
-                    unused_kwargs: Vec::new(),
-                }),
-        );
-    *added = true;
-}
-
-fn extend_check_result(target: &mut CheckResult, source: CheckResult) {
-    target.checked_kinds.extend(source.checked_kinds);
-    target.diagnostics.extend(source.diagnostics);
-}
-
-#[derive(Debug, Clone)]
-struct ExpandedChecks {
-    checks: Vec<CheckKind>,
-}
-
-fn expand_check_kinds(checks: Vec<CheckKind>) -> ExpandedChecks {
-    let defaults = vec![
-        CheckKind::Syntax,
-        CheckKind::References,
-        CheckKind::Untranslated,
-        CheckKind::Missing,
-        CheckKind::Stale,
-        CheckKind::Kwargs,
-    ];
-
-    let is_default_or_all = checks.is_empty() || checks.contains(&CheckKind::All);
-    let checks = if is_default_or_all {
-        defaults
-    } else {
-        normalize_check_order(checks)
-    };
-
-    let mut expanded = Vec::new();
-    for check in checks {
-        if !expanded.contains(&check) {
-            expanded.push(check);
-        }
-    }
-    ExpandedChecks { checks: expanded }
-}
-
-fn normalize_check_order(checks: Vec<CheckKind>) -> Vec<CheckKind> {
-    if !checks.contains(&CheckKind::Syntax) {
-        return checks;
-    }
-
-    let mut normalized = vec![CheckKind::Syntax];
-    normalized.extend(
-        checks
-            .into_iter()
-            .filter(|check| *check != CheckKind::Syntax),
-    );
-    normalized
-}
-
-fn has_fatal_syntax_diagnostics(result: &CheckResult) -> bool {
-    result.diagnostics.iter().any(|diagnostic| {
-        diagnostic.kind == DiagnosticKind::Syntax && diagnostic.severity == Severity::Error
-    })
-}
-
-fn dedup_diagnostics(result: &mut CheckResult) {
-    let mut seen = FastHashSet::default();
-    result.diagnostics.retain(|diagnostic| {
-        diagnostic.kind != DiagnosticKind::Extraction || seen.insert(diagnostic_key(diagnostic))
-    });
-}
-
-fn diagnostic_key(diagnostic: &Diagnostic) -> String {
-    format!(
-        "{}|{}|{}|{}|{}|{}|{}|{}|{}",
-        diagnostic.kind.as_str(),
-        diagnostic.severity.as_str(),
-        diagnostic.locale.as_deref().unwrap_or(""),
-        diagnostic.key.as_deref().unwrap_or(""),
-        diagnostic.message,
-        location_key(diagnostic.ftl_location.as_ref()),
-        location_key(diagnostic.code_location.as_ref()),
-        diagnostic.missing_kwargs.join(","),
-        diagnostic.unused_kwargs.join(","),
-    )
-}
-
-fn location_key(location: Option<&check::SourceLocation>) -> String {
-    let Some(location) = location else {
-        return String::new();
-    };
-
-    let line = location.line.map_or(String::new(), |line| line.to_string());
-    let column = location
-        .column
-        .map_or(String::new(), |column| column.to_string());
-
-    format!("{}:{line}:{column}", location.path.display())
-}
-
-impl From<FailSeverity> for Severity {
-    fn from(value: FailSeverity) -> Self {
-        match value {
-            FailSeverity::Error => Self::Error,
-            FailSeverity::Warn => Self::Warn,
-        }
-    }
-}
-
-fn normalize_output_path(path: PathBuf, format: &CheckOutputFormat) -> PathBuf {
-    if path.extension().is_some() {
-        return path;
-    }
-
-    let suffix = match format {
-        CheckOutputFormat::Terminal => "txt",
-        CheckOutputFormat::Json => "json",
-    };
-
-    path.with_extension(suffix)
-}
-
-fn cli_or_config_vec<T>(cli: Vec<T>, config: Option<Vec<T>>, default: Vec<T>) -> Vec<T> {
-    if !cli.is_empty() {
-        cli
-    } else if let Some(config) = config {
-        config
-    } else {
-        default
-    }
-}
-
-fn cli_or_config_path(
-    cli: Option<PathBuf>,
-    config: Option<PathBuf>,
-    base_dir: &Path,
-) -> Option<PathBuf> {
-    cli.or_else(|| resolve_config_path(config, base_dir))
-}
-
-fn resolve_required_path(
-    cli: Option<PathBuf>,
-    config: Option<PathBuf>,
-    base_dir: &Path,
-    error: &'static str,
-) -> Result<PathBuf> {
-    cli_or_config_path(cli, config, base_dir).context(error)
-}
-
-fn cli_or_config_enum<T>(cli: Option<T>, config: Option<String>, field: &str) -> Result<Option<T>>
-where
-    T: ValueEnum,
-{
-    if cli.is_some() {
-        return Ok(cli);
-    }
-
-    let Some(config) = config else {
-        return Ok(None);
-    };
-
-    T::from_str(&config, true).map(Some).map_err(|_| {
-        let values = T::value_variants()
-            .iter()
-            .filter_map(|variant| variant.to_possible_value())
-            .map(|value| value.get_name().to_string())
-            .collect::<Vec<_>>()
-            .join(", ");
-        anyhow::anyhow!("Invalid `{field}` value `{config}`. Expected one of: {values}")
-    })
-}
-
-fn cli_or_config_enum_vec<T>(
-    cli: Vec<T>,
-    config: Option<Vec<String>>,
-    field: &str,
-    default: Vec<T>,
-) -> Result<Vec<T>>
-where
-    T: ValueEnum,
-{
-    if !cli.is_empty() {
-        return Ok(cli);
-    }
-
-    let Some(config) = config else {
-        return Ok(default);
-    };
-
-    config
-        .into_iter()
-        .map(|value| {
-            T::from_str(&value, true).map_err(|_| {
-                let values = T::value_variants()
-                    .iter()
-                    .filter_map(|variant| variant.to_possible_value())
-                    .map(|value| value.get_name().to_string())
-                    .collect::<Vec<_>>()
-                    .join(", ");
-                anyhow::anyhow!("Invalid `{field}` value `{value}`. Expected one of: {values}")
-            })
-        })
-        .collect()
-}
-
-fn exit_config_error(error: anyhow::Error) -> ! {
-    error!(target: "cli", "Configuration error: {}", error);
-    std::process::exit(2);
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use extractor::ftl::diagnostics::{
-        CodeLocation, ExtractionDiagnostic, ExtractionDiagnosticKind,
-    };
-    use tempfile::TempDir;
-
-    #[test]
-    fn check_all_runs_syntax_first() {
-        let expanded = expand_check_kinds(vec![CheckKind::All]);
-
-        assert_eq!(expanded.checks.first(), Some(&CheckKind::Syntax));
-        assert_eq!(
-            expanded.checks,
-            vec![
-                CheckKind::Syntax,
-                CheckKind::References,
-                CheckKind::Untranslated,
-                CheckKind::Missing,
-                CheckKind::Stale,
-                CheckKind::Kwargs,
-            ]
-        );
-    }
-
-    #[test]
-    fn check_custom_list_is_not_default_or_all() {
-        let expanded = expand_check_kinds(vec![CheckKind::Missing, CheckKind::Syntax]);
-
-        assert_eq!(expanded.checks, vec![CheckKind::Syntax, CheckKind::Missing]);
-    }
-
-    #[test]
-    fn custom_checks_move_syntax_first() {
-        let expanded = expand_check_kinds(vec![
-            CheckKind::Kwargs,
-            CheckKind::Missing,
-            CheckKind::References,
-            CheckKind::Stale,
-            CheckKind::Syntax,
-        ]);
-
-        assert_eq!(
-            expanded.checks,
-            vec![
-                CheckKind::Syntax,
-                CheckKind::Kwargs,
-                CheckKind::Missing,
-                CheckKind::References,
-                CheckKind::Stale,
-            ]
-        );
-    }
-
-    #[test]
-    fn custom_checks_without_syntax_keep_order() {
-        let expanded = expand_check_kinds(vec![CheckKind::Missing, CheckKind::Kwargs]);
-
-        assert_eq!(expanded.checks, vec![CheckKind::Missing, CheckKind::Kwargs]);
-    }
-
-    #[test]
-    fn combined_code_aware_checks_extract_once() {
-        let temp = TempDir::new().unwrap();
-        let locales_path = temp.path().join("locales");
-        let code_path = temp.path().join("code");
-        std::fs::create_dir_all(locales_path.join("uk")).unwrap();
-        std::fs::create_dir_all(&code_path).unwrap();
-        std::fs::write(locales_path.join("uk").join("_default.ftl"), "").unwrap();
-
-        let expanded_checks = expand_check_kinds(vec![
-            CheckKind::Missing,
-            CheckKind::Stale,
-            CheckKind::Kwargs,
-        ]);
-        let config = CheckRunConfig {
-            locales_path,
-            code_path: Some(code_path.clone()),
-            locales: vec!["uk".to_string()],
-            suggest_from: Vec::new(),
-            i18n_keys: FastHashSet::default(),
-            i18n_keys_prefix: FastHashSet::default(),
-            exclude_dirs: FastHashSet::default(),
-            ignore_attributes: FastHashSet::default(),
-            ignore_kwargs: FastHashSet::default(),
-            default_ftl_file: PathBuf::from(DEFAULT_FTL_FILENAME),
-            cache: false,
-            cache_path: None,
-            clear_cache: false,
-        };
-
-        let mut extraction_calls = 0;
-        let result = run_check_with_extractor(expanded_checks, config, |_| {
-            extraction_calls += 1;
-            Ok(ExtractedCode {
-                keys: Vec::new(),
-                diagnostics: vec![ExtractionDiagnostic {
-                    kind: ExtractionDiagnosticKind::KeyPathConflict,
-                    key: "hello".to_string(),
-                    message: "Fluent key hello has different paths".to_string(),
-                    locations: vec![CodeLocation {
-                        path: code_path.join("app.py"),
-                        line: 1,
-                        column: 1,
-                    }],
-                }],
-                py_files_count: 1,
-            })
-        })
-        .unwrap();
-
-        assert_eq!(extraction_calls, 1);
-        assert_eq!(
-            result
-                .diagnostics
-                .iter()
-                .filter(|diagnostic| diagnostic.kind == DiagnosticKind::Extraction)
-                .count(),
-            1
-        );
     }
 }
